requirement.txt
alembic==1.0.10
blinker==1.4
Bootstrap-Flask==1.0.10
certifi==2019.3.9
chardet==3.0.4
Click==7.0
dominate==2.3.5
elasticsearch==7.0.2
Flask==1.0.3
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Migrate==2.5.2
Flask-Moment==0.7.0
Flask-SQLAlchemy==2.4.0
Flask-WTF==0.14.2
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
langdetect==1.0.7
Mako==1.0.10
MarkupSafe==1.1.1
PyJWT==1.7.1
python-dateutil==2.8.0
python-editor==1.0.4
requests==2.22.0
six==1.12.0
SQLAlchemy==1.3.4
urllib3==1.25.3
visitor==0.1.3
Werkzeug==0.15.4
WTForms==2.2.1
